pyproject.toml
[project]
name = "ralph-loop"
version = "0.2.0"
description = "Autonomous coding agent - task to spec to code with QA, using Claude Agent SDK or Deep Agents SDK"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    # Agent SDKs
    "claude-agent-sdk>=0.1.50",
    "deepagents>=0.4.10",

    # Core
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.scripts]
ralph = "ralph.cli:cli"

[project.optional-dependencies]
web = [
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "watchfiles>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ralph"]
